==> notebooks/0_descripcion_dataset.ipynb <==
# Descripción del dataset

## Contexto del conjunto de datos

El dataset que se utilizará para el estudio de esta práctica se extrae del conjunto de datos **Spanish Houses**, construido en la 1.práctica de la asignatura mediante técnicas de Web Scraping (*el trabajo puede encontrarse en el siguiente repositorio: https://github.com/trueuoc/spa_housing_crawler*)

En dicha práctica se implementó un rastreador web con el cual se consiguieron recoger 100.000 anuncios de viviendas en venta y alquiler del portal *idealistas.com*. Estas viviendas correspondían con viviendas de diversas provincias y comunidades del terrotorio español. Para el análisis de esta práctica, sin embargo, en vez de utilizar todas estas viviendas, se opta por acotar su ubicación, y emplear las viviendas de tan solo 2 provincias: las de **Bizkaia** y **Gipuzkoa**.

## Motivación del estudio

La motivación que precede a este estudio, corresponde, principalmente, con un un afán por comprender algo mejor los entresijos y factores que influyen en el precio de las viviendas. 

Y es que, a la hora de comprar o vender una casa, a menudo nos encontramos con que no tenemos muy claro si estamos pagando o pidiendo poco o mucho dinero. Tenemos referencias para otra clase de productos más cotidianos: sabemos que pagar 8 euros por una barra de pan es excesivo, y que ir a comer un menú del día por 2 euros es sospechosamente poco. Sin embargo, cuando se trata del precio de las viviendas, tratamos con cifras tan elevadas que ya no sabemos si estamos pagando poco o mucho, o si deberíamos de pedir algo más por nuestra vivienda de lo que actualmente estamos pidiendo.

Por esta razón, ante este desconocimiento acabamos por fiar todo lo relacionado con este tipo de cuestiones a inmobiliarias y asesorías que, la mayoría de veces, además de sustraer una comisión considerable, no tienen por principal objetivo beneficiar a sus clientes/consumidores.

Por ello, creemos que, realizar un análisis del precio de las viviendas puede servirnos para, al menos, empezar a entender mejor por qué las viviendas tienen el precio que tienen, y así dotarnos de cierto criterio para evaluar si una vivienda está sobrestimada o infravalorada.

## Objetivos del estudio

En este sentido, con el presente proyecto se establecen los siguientes objetivos principales:

+ Integrar y limpiar el conjunto de datos que se pretende estudiar
+ Realizar un análisis exploratorio de los datos que nos permita extraer conocimiento, enfatizando sobre todo en el estudio de relaciones entre las diferentes características de las viviendas y su precio
+ Implementar dos modelos de regresión
    + uno para ayudarnos a descubrir si una vivienda se ajusta a nuestro presupuesto
    + y otro para tratar de predecir el precio de las viviendas a partir de sus características
