==> notebooks/1_integracion_datos.ipynb <==
Cargamos las librerías que emplearemos durante el proceso:

import pandas as pd
import pandas_profiling

# Ampliamos el límite de columnas visibles para poder visualizarlas todas
pd.set_option('display.max_columns', 500)

# Categorización de los datos

En este apartado realizaremos una breve descripción de los datos que utilizaremos en este estudio.

Así, mayormente emplearemos un conjunto de datos que contiene información sobre anuncios de casas en venta de las provincias de Bizkaia y Gipuzkoa. Para ello, disponemos de 2 archivos, uno por cada provincia. Por lo tanto, cargamos los datos y los combinamos en un solo dataframe:

houses_biz = pd.read_csv('./raw_data/houses_vizcaya.csv')
houses_gip = pd.read_csv('./raw_data/houses_guipuzcoa.csv')

houses = pd.concat([houses_biz, houses_gip])

Una vez tenemos nuestro dataset, veamos de cuántas casas y variables consta:

print(f'Número de casas: {houses.shape[0]}')
print(f'Número de variables por casa: {houses.shape[1]}')

Asimismo, describamos cómo son estas variables:

**PROPIEDADES DEL ANUNCIO**

+ [0] __<font color=#5F66A1>house_id</font>__: Identificador numérico único del anuncio de la vivienda (coincide con el path del anuncio) &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ [1] __<font color=#5F66A1>ad_last_update</font>__: Fecha de la última actualización del anuncio &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font>
+ [2] __<font color=#5F66A1>ad_description</font>__: Descripción de texto del anuncio de la venta/alquiler de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font>

**PROPIEDADES DE LA VIVIENDA**

+ [3] __<font color=#5F66A1>price</font>__: Precio de alquiler/venta de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ [4] __<font color=#5F66A1>bath_num</font>__: Número de baños de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font> 
+ [5] __<font color=#5F66A1>condition</font>__: Estado en el que se encuentra la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat}</font>
+ [6] __<font color=#5F66A1>construct_date</font>__: Año de construcción de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#ED553B>{date}</font>
+ [7] __<font color=#5F66A1>energetic_certif</font>__: Certificado energético de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat}</font>
+ [8] __<font color=#5F66A1>floor</font>__: Número y/o ubicación de la(s) planta(s) de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat}</font>
+ [9] __<font color=#5F66A1>ground_size</font>__: Metros cuadrados del terreno donde se ubica la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ [10] __<font color=#5F66A1>heating</font>__: Sistema de calefacción de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat}</font>
+ [11] __<font color=#5F66A1>house_type</font>__: Tipo de vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat}</font>
+ [12] __<font color=#5F66A1>m2_real</font>__: Metros cuadrados totales de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ [13] __<font color=#5F66A1>m2_useful</font>__: Metros cuadrados útiles de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ [14] __<font color=#5F66A1>orientation</font>__: Orientación de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat}</font>
+ [15] __<font color=#5F66A1>room_num</font>__: Número de habitaciones de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>

**EQUIPAMIENTO ADICIONAL DE LA VIVIENDA**

+ [16] __<font color=#5F66A1>air_conditioner</font>__: Indica si la vivienda posee aire acondicionado &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [17] __<font color=#5F66A1>balcony</font>__: Indica si la vivienda tiene balcones &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [18] __<font color=#5F66A1>built_in_wardrobe</font>__: Indica si la vivienda consta de armarios empotrados &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [19] __<font color=#5F66A1>chimney</font>__: Indica si la vivienda consta de chimenea &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [20] __<font color=#5F66A1>garage</font>__: Indica el precio del garaje (en caso de que tenga) &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ [21] __<font color=#5F66A1>garden</font>__: Indica si la vivienda tiene jardín &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [22] __<font color=#5F66A1>kitchen</font>__: Indica si la vivienda está ya equipada con cocina &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [23] __<font color=#5F66A1>lift</font>__: Indica si la vivienda consta de ascensor &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [24] __<font color=#5F66A1>reduced_mobility</font>__: Indica si la vivienda está adaptada para personas con movilidad reducida &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [25] __<font color=#5F66A1>storage_room</font>__: Indica si la vivienda consta de trastero &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [26] __<font color=#5F66A1>swimming_pool</font>__: Indica si la vivienda tiene piscina &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [27] __<font color=#5F66A1>terrace</font>__: Indica si la vivienda tiene terraza &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>
+ [28] __<font color=#5F66A1>unfurnished</font>__: Indica si la vivienda está sin amueblar &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{cat binary}</font>

**UBICACIÓN DE LA VIVIENDA**

+ [29] __<font color=#5F66A1>loc_full</font>__: Dirección completa de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font>
+ [30] __<font color=#5F66A1>loc_zone</font>__: Provincia en la que se ubica la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font>
+ [31] __<font color=#5F66A1>loc_district</font>__: Área de la provincia en la que se ubica la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font>
+ [32] __<font color=#5F66A1>loc_city</font>__: Localidad en la que se ubica la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font>
+ [33] __<font color=#5F66A1>loc_neigh</font>__: Vecindario en la que se ubica la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font>
+ [34] __<font color=#5F66A1>loc_street</font>__: Calle en la que se ubica la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#f3ca75>{text}</font> 

**METADATOS**
+ [35] __<font color=#5F66A1>obtention_date</font>__: Fecha de obtención de los datos de la vivienda &nbsp;&nbsp;&nbsp;&nbsp;<font color=#599d70>{date}</font>

<br />

# Enriquecimiento de los datos

En este apartado enriqueceremos nuestro conjunto de datos principal añadiendo información demográfica y socio-económica. Concretamente las variables que añadiremos serán la siguiente:

+ __<font color=#5F66A1>pop_total</font>__: Número de habitantes por municipio &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ Distribución de edades por municipio 
    + __<font color=#5F66A1>pop_0-19</font>__: Tasa de jovenes menores de 20 años &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
    + __<font color=#5F66A1>pop_20-64</font>__: Tasa de adultos entre 20 y  64 años &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
    + __<font color=#5F66A1>pop_>=64</font>__: Tasa de adultos mayores de 64 años &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>
+ __<font color=#5F66A1>pop_foreign</font>__: Tasa de extranjeros por municipio &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num float}</font>
+ __<font color=#5F66A1>rent</font>__: Renta media personal por municipio &nbsp;&nbsp;&nbsp;&nbsp;<font color=#c874b9>{num}</font>

## Integración de información demográfica

Empezaremos añadir el número de habitantes que viven en cada municipio. Para ello, empezamos cargando y explorando los datos demográficos:

population = pd.read_csv('./raw_data/pop2018.csv', sep=';')
population.head(3)

Vemos que el nombre de cada municipio, esta precedido por el código postal. Lo eliminamos:

population['city'] = population['city'].apply(lambda s: ' '.join(s.split(' ')[1:]))
population.head(3)

Añadimos los datos demográficos de cada municipio a nuestro conjunto de datos, utilizando para ello los nombres de los municipios:

houses_pop = pd.merge(left=houses, right=population, how='left', left_on='loc_city', right_on='city')

Comprobamos que no haya ningún municipio al que no se le haya asignado el número de habitantes:

print(houses_pop
        .loc[houses_pop['city'].isna()]['loc_city']
        .unique())

Vemos que hay unos cuantos municipios que no han sido conciliados. Los corregimos de forma manual, combinamos, y volvemos a comprobar:

houses = houses.replace({'loc_city':
              {
                  'Carranza':'Karrantza Harana/Valle de Carranza',
                  'Arcentales':'Artzentales',
                  'Trucios':'Trucios-Turtzioz',
                  'Guizaburuaga':'Gizaburuaga',
                  'Amorebieta-Echano':'Amorebieta-Etxano',
                  'Izurza':'Izurtza',
                  'Valle de Trapaga-Trapagaran':'Valle de Trápaga-Trapagaran',
                  'Astrabudua':'Erandio',
                  'Etxebarri, Anteiglesia de San esteban':'Etxebarri',
                  'Alonsotegui':'Alonsotegi',       
                  'Abanto y Ciérvana/Abanto Zierbena':'Abanto y Ciérvana-Abanto Zierbena',
                  'Gallarta':'Abanto y Ciérvana-Abanto Zierbena',
                  'Gauteguiz de Arteaga':'Gautegiz Arteaga',
                  'Ibarranguelua':'Ibarrangelu',
                  'Orduña':'Urduña/Orduña',
                  'Ubidea':'Ubide',
                  'Aracaldo':'Arakaldo',
                  'Maruri':'Maruri-Jatabe',
                  'Sopelana':'Sopela',                 
                  'Bidegoyan':'Bidania-Goiatz',
                  'Leaburu-Gaztelu':'Leaburu',
                  'Donostia-San Sebastián':'Donostia/San Sebastián',
                  'Gatica':'Gatika',
                  'Arrasate o Mondragon':'Arrasate/Mondragón',
                  'Renteria / Errenteria':'Errenteria',
                  'Legazpia':'Legazpi',
                  'Gainza':'Gaintza',
                  'Altzola':'Elgoibar',
                  'Soraluze/Placencia de las Armas':'Soraluze-Placencia de las Armas'
              }});

houses_pop = pd.merge(left=houses, right=population, how='left', left_on='loc_city', right_on='city')
print(houses_pop
        .loc[houses_pop['city'].isna()]['loc_city']
        .unique())

A continuacion, añadimos la información sobre la distribución de grupos de edades de cada población. Para ello, empezamos cargando y explorando los datos relacionados:

pop_age = pd.read_csv('./raw_data/pop2018_age.csv', sep=';')
pop_age.head(6)

Vemos que, por cada municipio, tenemos 3 grupos de edades diferenciados. No obstante, cada grupo debería de tener su columna, por lo que procedemos a reposicionar los datos de mejor forma:

pop_age = pd.pivot_table(pop_age,
                         index='city',
                         columns='pop_age',
                         values='percent')

pop_age.head(3)

Cambiamos los nombres de las columnas para que sean más significativos, y añadimos la información de la edad al dataset principal, usando nuevamente el municipio, y comprobando que todos los nombres se hayan conciliado bien:

pop_age.rename(columns={'0 - 19':'pop_0-19','20 - 64':'pop_20-64',
                          '>= 65':'pop_>=65'}, inplace=True)

houses_pop = pd.merge(left= houses_pop, right=pop_age,  how='left', on='city')
print(houses_pop
        .loc[houses_pop['pop_0-19'].isna()]['loc_city']
        .unique())

Vemos que, nuevamente, hay municipios que no se han conciliado de forma correcta. Los corregimos:

pop_age = pop_age.rename(index=
              {
                  'Sopelana':'Sopela',
                  'Donostia / San Sebastián':'Donostia/San Sebastián',
                  'Soraluze/Placencia de las Armas':'Soraluze-Placencia de las Armas'
              });

houses_pop = houses_pop.drop(columns=['pop_0-19', 'pop_20-64', 'pop_>=65'])           
houses_pop = pd.merge(left=houses_pop, right=pop_age, how='left', on='city')

print(houses_pop
        .loc[houses_pop['pop_0-19'].isna()]['loc_city']
        .unique())

Tras incluir la información sobre edades, procedemos ahora a añadir la tasa de extranjeros por cada población. Para ello, nuevamente, empezamos cargando y explorando el archivo correspondiente:

pop_for = pd.read_csv('./raw_data/pop2018_foreign.csv', sep=';')
pop_for.head(3)

Observamos que disponemos ya del porcentaje de extranjeros por municipio. Al no requerir ningún tratamiento, añadimos directamente la información al conjunto de datos principal:

houses_pop = pd.merge(left=houses_pop, right=pop_for, how='left', on='city')
print(houses_pop
        .loc[houses_pop['pop_foreign'].isna()]['loc_city']
        .unique())

Otra vez, corregimos la conciliación manualmente:

pop_for = pop_for.replace({'city':
              {
                  'Sopelana':'Sopela',
                  'Donostia / San Sebastián':'Donostia/San Sebastián',
                  'Soraluze/Placencia de las Armas':'Soraluze-Placencia de las Armas'
              }});

houses_pop = houses_pop.drop(columns=['pop_foreign'])           
houses_pop = pd.merge(left=houses_pop, right=pop_for, how='left', on='city')
print(houses_pop
        .loc[houses_pop['pop_foreign'].isna()]['loc_city']
        .unique())

##  Integración de información socio-económica

Una vez hemos añadido los datos demográficos, haremos lo propio con la renta personal media. Empezaremos por cargar y explorar el archivo de datos de las rentas:

rent = pd.read_csv('./raw_data/rent2016.csv', sep=';')
rent.head(10)

Vemos que por cada municipio tenemos varios tipos de renta. Seleccionamos la renta total y descartamos el resto: 

rent = (rent.loc[rent.rent_type=='Renta total']
                .drop(columns='rent_type'))
rent.head(3)

Finalmente, añadimos la información de renta personal por municipio al conjunto de datos principal y comprobamos si se han conciliado bien los nombres:

houses_pop_rent = pd.merge(left=houses_pop, right=rent, how='left', on='city')

print(houses_pop_rent
        .loc[houses_pop_rent['rent'].isna()]['loc_city']
        .unique()
)

Vemos que hay municipios que no se han conciliado, por lo que los corregimos de forma manual:

rent = rent.replace({'city':
              {
                  'Urdúliz':'Urduliz',
                  'Bérriz':'Berriz',
                  'Donostia / San Sebastián':'Donostia/San Sebastián',
                  'Soraluze/Placencia de las Armas':'Soraluze-Placencia de las Armas'
              }});

houses_pop_rent = pd.merge(left=houses_pop, right=rent, how='left', on='city')

print(houses_pop_rent
        .loc[houses_pop_rent['rent'].isna()]['loc_city']
        .unique()
)

# Preselección de variables

En este apartado realizaremos una preselección de las variables que emplearemos a lo largo del estudio, descartando aquellas que sabemos de antemano que no nos serán de utilidad.

Las variables que se descartarán son las siguientes:

+ __<font color=#5F66A1>ad_last_update</font>__
+ __<font color=#5F66A1>ad_description</font>__
+ __<font color=#5F66A1>loc_full</font>__
+ __<font color=#5F66A1>loc_street</font>__
+ __<font color=#5F66A1>obtention_date</font>__

houses = houses_pop_rent.reindex(columns=['house_id', 'ad_description', 'price', 'bath_num'
                                , 'condition', 'construct_date', 'energetic_certif', 'floor', 'ground_size'
                                , 'heating', 'house_type', 'm2_real', 'm2_useful', 'orientation', 'room_num'
                                , 'air_conditioner', 'balcony', 'built_in_wardrobe', 'chimney', 'garage'
                                , 'garden', 'kitchen', 'lift', 'reduced_mobility', 'storage_room'
                                , 'swimming_pool', 'terrace', 'unfurnished', 'loc_zone', 'loc_district', 'loc_city'
                                , 'loc_neigh', 'pop_total', 'pop_0-19', 'pop_20-64', 'pop_>=65', 'pop_foreign', 'rent'      
                                ])

Por lo tanto, nuestro dataset queda caracterizado de la siguiente manera:

print(f'Número de casas: {houses.shape[0]}')
print(f'Número de variables por casa: {houses.shape[1]}')

<br />

Por otro lado, realizamos una primera exploración descriptiva algo mas profunda, mostrando  el porcentaje de *missing values*, los valores más comunes, estadísticos principales, así como distribuciones, para lo cual, nos valdremos de *pandas_profiling*.

pandas_profiling.ProfileReport(houses)


Para concluir con la fase de integración, almacenamos nuestro conjunto de datos procesado:

houses.to_csv('../processed_data/houses_integrated.csv', index=False)
