==> notebooks/5_conclusiones.ipynb <==
# Conclusiones del análisis

A lo largo de este estudio hemos visto como, a grandes rasgos, partiendo de un conjunto de datos que contenía información sobre anuncios de viviendas en venta se ha implementado un proceso mediante el cual, en última instancia, se ha logrado extraer información valiosa relacionada con el precio de las viviendas.

<br />

El estudio se ha iniciado con la fase de **INTEGRACIÓN DE DATOS**. En esta primera fase se han caracterizado las distintas variables del conjunto de datos principal. Este conjunto de datos, además, se ha enriquecido mediante la adición de información demográfica y socio-económica. Para ello, ha sido necesario conciliar cuidadosamente los nombres de las localidades, ya que cada archivo podía contener los nombres de los municipios en diferentes idiomas.

La siguiente fase del estudio ha sido la **LIMPIEZA DE DATOS**. Para ello, el primer paso ha consistido en estandarizar las distintas variables, asignando los valores del modo más representativo posible a fin de mejorar su manejabilidad e interpretabilidad. En dicho proceso, además, se han generado nuevas variables, tales como *location*, o *floor*, como consecuencia de manipular o dividir otras ya existentes.

Tras la estandarización, se ha verificado que no hubiera viviendas con más de un anuncio. Para ello, el principal reto ha sido implementar un filtro que fuese capaz de detectar viviendas similares aun con anuncios que presentaran ciertas incoherencias y disimilitudes entre sí.

Por otro lado, se ha llevado a cabo también una gestión de *outliers*, en la cual se han explorado los valores extremos, evaluando en cada caso la veracidad de dichos valores. Por último, la fase de *limpieza de datos* se ha concluido con un estudio de los *missing values*, en el cual se han imputando los datos faltantes de diversas formas (por asignación manual, por moda, por similitud mediante k-NN, ...).

Una vez ya resultas las incosistencias de los datos, la tercera fase del estudio ha sido el **ANÁLISIS EXPLORATORIO DE LOS DATOS**. En esta fase, se han estudiado las distribuciones y relaciones de distintas variables, haciendo sobretodo hincapié en tratar de entender cómo y en qué medida afectan las distintas características de las viviendas en su precio. Para estudiar estas relaciones, principalmente se ha hecho uso de un gran número de visualizaciones, tales como *boxplots*, *barplots*, *countplots*, ... Además, estas visualizaciones se han apoyado con una serie de tests estadísticos, tales como test de normalidad de *Shapiro-Wilk*, test de correlación de *Kendall*, prueba de *Levene* o test de *Kruskal-Wallis*.

Finalmente, la última fase del estudio ha consistido en la **IMPLEMENTACIÓN DE MODELOS PREDICTIVOS**, concretamente 2. El primero de ellos ha sido una regresión con la cual, se ha buscado predecir si una vivienda cuesta más o menos que cierta cantidad. Así, con este modelo se ha querido simular un escenario en el cual, alguien, pudiera comprobar si puede o no encontrar una vivienda de ciertos parámetro. Por otro lado, incluso las inmobiliarias podrían beneficiarse de este modelo, pudiéndolo emplear para valorar junto al cliente si el precio que tienen pensado para la vivienda es coherente.

La precisión del modelo resultante ha sido del 91.47%, lo cual significa que ha sido capaz de predecir en más de 9 de cada 10 viviendas si ésta era mayor o menor que nuestro umbral de precio . 

El segundo modelo que se ha implementado ha un modelo de regresión que predice el precio de las viviendas a partir de sus características. Para este segundo modelo se ha utilizado la técnica XG-Boost. De este modo, el principal reto ha sido la búsqueda de una combinación de hiperparámetros que mejorara el rendimiento lo máximo posible. Así, finalmente, se ha conseguido un modelo con un error de predicción del 17,18%, capaz de predecir de forma acertada el precio de un 85% de las viviendas, con un margen de +/- 50.000€.

Además, se han visto cuáles son las variables más relevantes a la hora de predecir el precio, observando como el tamaño de las viviendas es la característica que más impacto tiene. Por otro lado, finalmente se ha estudiado si la inclusión de más datos en el estudio puede favorecer un mejor rendimiento del modelo, resolviéndose la questión en que sí.
